switch interpretation of :null temporarily
The data format is planned to switch to :1 to mean N, so digitMeansNull is added to enable working with the current test data, and switch this interpretation when the new data is received.

vcf-feature.js : addFeaturesJson() : add digitMeansNull = false, which flips the interpretation to : ( ./.:0 is null, ./.:1 is missing).
vcf-genotype.js : vcfGenotypeSamplesFiltered() : refToGenotype() : matchRef === null : match 0 instead of 1.  Also match dots exactly in ./. because . without escape could match [01].